+++ packages/zowe-explorer/__tests__/__unit__/uss/actions.unit.test.ts
@@ -661,61 +661,6 @@ describe("USS Action Unit Tests - Functions uploadDialog & uploadFile", () => {
     });
 });
 
-describe("USS Action Unit Tests - Function changeFileType", () => {
-    async function createBlockMocks(globalMocks) {
-        const newMocks = {
-            node: null,
-            testUSSTree: null,
-            getMvsApiMock: jest.fn(),
-            testResponse: createFileResponse({ etag: "132" }),
-            testDoc: createTextDocument(path.normalize("/sestest/tmp/foo.txt")),
-            ussNode: createUSSNode(globalMocks.testSession, createIProfile()),
-            mvsApi: ZoweExplorerApiRegister.getMvsApi(globalMocks.testProfile),
-            putUSSPayload: jest.fn().mockResolvedValue(`{"stdout":[""]}`),
-        };
-
-        newMocks.node = new ZoweUSSNode({
-            label: "u/myuser/testFile",
-            collapsibleState: vscode.TreeItemCollapsibleState.None,
-            parentNode: newMocks.ussNode,
-            parentPath: "/",
-        });
-        newMocks.ussNode.children.push(newMocks.node);
-        newMocks.testUSSTree = createUSSTree(
-            [createFavoriteUSSNode(globalMocks.testSession, globalMocks.testProfile)],
-            [newMocks.ussNode],
-            createTreeView()
-        );
-        globalMocks.ussFile.mockResolvedValueOnce(newMocks.testResponse);
-        globalMocks.withProgress.mockImplementation((progLocation, callback) => callback());
-        newMocks.getMvsApiMock.mockReturnValue(newMocks.mvsApi);
-        ZoweExplorerApiRegister.getMvsApi = newMocks.getMvsApiMock.bind(ZoweExplorerApiRegister);
-
-        Object.defineProperty(globalMocks.Utilities, "putUSSPayload", {
-            value: newMocks.putUSSPayload,
-            configurable: true,
-        });
-
-        return newMocks;
-    }
-
-    it("Tests that changeFileType() runs successfully", async () => {
-        const globalMocks = createGlobalMocks();
-        const blockMocks = await createBlockMocks(globalMocks);
-        const node = new ZoweUSSNode({ label: "node", collapsibleState: vscode.TreeItemCollapsibleState.None, parentNode: blockMocks.ussNode });
-        Object.defineProperty(node, "getUSSDocumentFilePath", { value: jest.fn().mockReturnValueOnce(blockMocks.testDoc), configurable: true });
-        Object.defineProperty(fs, "existsSync", { value: jest.fn().mockReturnValueOnce(false), configurable: true });
-
-        node.binary = true;
-        node.contextValue = globals.USS_BINARY_FILE_CONTEXT;
-        node.getSessionNode().encodingMap[node.fullPath] = { kind: "binary" };
-        expect(node.binary).toBeTruthy();
-
-        await ussNodeActions.changeFileType(node, false, blockMocks.testUSSTree);
-        expect(node.binary).toBeFalsy();
-    });
-});
-
 describe("USS Action Unit Tests - function uploadFile", () => {
     async function createBlockMocks(globalMocks) {
         const newMocks = {