+++ packages/zowe-explorer/__tests__/__unit__/uss/init.unit.test.ts
@@ -123,20 +123,6 @@ describe("Test src/dataset/extension", () => {
                 ],
             },
             {
-                name: "zowe.uss.binary",
-                mock: [
-                    { spy: jest.spyOn(contextuals, "isText"), arg: [test.value], ret: true },
-                    { spy: jest.spyOn(ussActions, "changeFileType"), arg: [test.value, true, ussFileProvider] },
-                ],
-            },
-            {
-                name: "zowe.uss.text",
-                mock: [
-                    { spy: jest.spyOn(contextuals, "isBinary"), arg: [test.value], ret: true },
-                    { spy: jest.spyOn(ussActions, "changeFileType"), arg: [test.value, false, ussFileProvider] },
-                ],
-            },
-            {
                 name: "zowe.uss.renameNode",
                 mock: [{ spy: jest.spyOn(ussFileProvider, "rename"), arg: [test.value] }],
             },
@@ -207,7 +193,7 @@ describe("Test src/dataset/extension", () => {
             },
             {
                 name: "zowe.uss.openWithEncoding",
-                mock: [{ spy: jest.spyOn(ussFileProvider, "openWithEncoding"), arg: [test.value] }],
+                mock: [{ spy: jest.spyOn(ussFileProvider, "openWithEncoding"), arg: [test.value, undefined] }],
             },
             {
                 name: "onDidChangeConfiguration",