+++ packages/zowe-explorer-api/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to the "zowe-explorer-api" extension will be documented in this file.
 
+## Recent Changes
+
+### New features and enhancements
+
+- **Breaking:** Removed the following properties/methods:
+  - `IZoweUSSTreeNode.binaryFiles` -> `IZoweUSSTreeNode.encodingMap`
+  - `IZoweUSSTreeNode.mProfileName` -> `IZoweUSSTreeNode.getProfileName()`
+  - `IZoweUSSTreeNode.setBinary()` -> `IZoweUSSTreeNode.setEncoding()`
+- **Breaking:** Removed `ZoweTreeNode.binary`, `ZoweTreeNode.binaryFiles`, and `ZoweTreeNode.shortLabel`. These properties are not applicable for all tree nodes and should be defined in subclasses of `ZoweTreeNode` if necessary.
+
 ## `3.0.0-next.202401241448`
 
 ### New features and enhancements